Extract moving by a certain amount out of the player
Instead, just set the position on the player, and calculate the new position based on a delta in the Game object. I think this will make it easier, later, to detect when a move should be allowed or not, since we will have to interact with the Map to know that.