Add function to get a single tile from the Map
This will let the Game object checking that a move is valid before letting the Player move.

One thing to note here is that "getTile" isn't going to be the most efficient algorithm. Right now it searches through each item in the tiles array looking for one matching the provided position. While it does stop as soon as it finds one, it would be more efficient to store the map height and width, and calculate the index from the provided position, and index the array directly. However, until the current implementation proves to be too slow, Im going to go with it because it's simpler to implement, and I'm lazy.